+++ NEWS.md
@@ -1,4 +1,9 @@
-# JSmediation (Development version)
+# JSmediation (development version)
+
+* Introduce `standardize_variables` to easily center and reduce your variables.
+* Introduce the `compute_indirect_effect_for()` function which computes the 
+  indirect effect confidence interval for a specific value of the moderator in 
+  a moderated mediation model.
 
 # JSmediation 0.1.2